requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
firebase-admin>=6.5.0
groq>=0.5.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
python-pptx>=0.6.23
Pillow>=10.2.0
pytesseract>=0.3.10
gunicorn>=21.2.0
